AnimatiApp Django/AnimatiBack/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-db-connection-pool==1.2.5
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
greenlet==3.1.1
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
setuptools==74.1.2
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
